+++ pmd-core/src/test/java/net/sourceforge/pmd/lang/DummyLanguageModule.java
@@ -31,14 +31,14 @@ public class DummyLanguageModule extends BaseLanguageModule {
 
     public DummyLanguageModule() {
         super(NAME, null, TERSE_NAME, DummyRuleChainVisitor.class, "dummy");
-        addVersion("1.0", new Handler(), true);
+        addVersion("1.0", new Handler(), false);
         addVersion("1.1", new Handler(), false);
         addVersion("1.2", new Handler(), false);
         addVersion("1.3", new Handler(), false);
         addVersion("1.4", new Handler(), false);
         addVersion("1.5", new Handler(), false);
         addVersion("1.6", new Handler(), false);
-        addVersion("1.7", new Handler(), false);
+        addVersion("1.7", new Handler(), true);
         addVersion("1.8", new Handler(), false);
     }