+++ src/site/markdown/overview/changelog.md
@@ -125,6 +125,7 @@ you'll need a java8 runtime environment.
     *   [#1458](https://sourceforge.net/p/pmd/bugs/1458/): Performance degradation scanning large XML files with XPath custom rules
     *   [#1461](https://sourceforge.net/p/pmd/bugs/1461/): Possible threading issue due to PR#75
     *   [#1470](https://sourceforge.net/p/pmd/bugs/1470/): Error with type-bound lambda
+    *   [#1478](https://sourceforge.net/p/pmd/bugs/1478/): PMD CLI - Use first language as default if Java is not available
 
 **API Changes:**